Split the ButtonsComponent into two classes
Each of these classes is simpler than the one
