+++ app/components/collections/buttons_component.html.erb
@@ -1,0 +1,12 @@
+<div class="row justify-content-between">
+  <div class="col-auto">
+    <%= delete_button %>
+  </div>
+  <div class="col-auto">
+    <%= cancel_button %>
+  </div>
+  <div class="col-auto">
+    <%= form.submit 'Save as draft', class: 'btn btn-primary', id: 'save-draft-button' %>
+    <%= form.submit 'Deposit', class: 'btn btn-primary' %>
+  </div>
+</div>
