+++ app/components/collections/form_component.html.erb
@@ -180,5 +180,5 @@
     </div>
   </section>
 
-  <%= render ButtonsComponent.new(form: form) %>
+  <%= render Collections::ButtonsComponent.new(form: form) %>
 <% end %>
