+++ app/components/works/form_component.html.erb
@@ -29,5 +29,5 @@
   <%= render Works::EmbargoComponent.new(form: f) %>
   <%= render Works::LicenseComponent.new(form: f) %>
   <%= render Works::AgreementComponent.new(form: f) %>
-  <%= render ButtonsComponent.new(form: f) %>
+  <%= render Works::ButtonsComponent.new(form: f) %>
 <% end %>
