+++ spec/components/works/buttons_component_spec.rb
@@ -3,7 +3,7 @@
 
 require 'rails_helper'
 
-RSpec.describe ButtonsComponent do
+RSpec.describe Works::ButtonsComponent do
   let(:component) { described_class.new(form: form) }
   let(:form) { ActionView::Helpers::FormBuilder.new(nil, work_form, controller.view_context, {}) }
   let(:work) { build_stubbed(:work, collection: build(:collection, id: 7)) }
